head2toe/requirements.txt
absl-py
numpy
ml-collections
sklearn
tensorflow==2.7
-e git+https://github.com/google-research/task_adaptation.git#egg=task_adaptation
